requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
SQLAlchemy==1.4.53
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
email-validator==2.1.0
marshmallow==3.20.1
python-dotenv==1.0.0